Moving cibyl utils to kernel
