+++ cibyl/outputs/cli/ci/system/impls/zuul/colored/cascades/hierarchy.py
@@ -17,7 +17,7 @@ from anytree import AsciiStyle, RenderTree
 
 from cibyl.models.ci.zuul.job import Job
 from cibyl.outputs.cli.printer import ColoredPrinter
-from cibyl.utils.strings import IndentedTextBuilder
+from kernel.tools.text import IndentedTextBuilder
 from kernel.tools.tree import Tree
