+++ cibyl/sources/zuul/apis/__init__.py
@@ -23,7 +23,7 @@ from cibyl.exceptions.source import SourceException
 from cibyl.sources.zuul.apis.providers import JobsProvider, PipelinesProvider
 from cibyl.sources.zuul.utils.artifacts import Artifact, ArtifactKind
 from cibyl.sources.zuul.utils.tests.types import TestSuite
-from cibyl.utils.io import Closeable
+from kernel.tools.io import Closeable
 
 
 class ZuulAPIError(SourceException):
