+++ cibyl/sources/zuul/utils/artifacts/manifest.py
@@ -22,7 +22,7 @@ from cibyl.sources.zuul.apis import Artifact, ArtifactKind
 from cibyl.sources.zuul.apis.http import ZuulHTTPBuildAPI as Build
 from cibyl.sources.zuul.utils.artifacts import ArtifactError
 from cibyl.utils.filtering import matches_regex
-from cibyl.utils.net import download_into_memory
+from kernel.tools.net import download_into_memory
 
 
 @dataclass
