--- cibyl/utils/io.py
@@ -1,27 +1,0 @@
-"""
-#    Copyright 2022 Red Hat
-#
-#    Licensed under the Apache License, Version 2.0 (the "License"); you may
-#    not use this file except in compliance with the License. You may obtain
-#    a copy of the License at
-#
-#         http://www.apache.org/licenses/LICENSE-2.0
-#
-#    Unless required by applicable law or agreed to in writing, software
-#    distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
-#    WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
-#    License for the specific language governing permissions and limitations
-#    under the License.
-"""
-from abc import ABC, abstractmethod
-
-
-class Closeable(ABC):
-    """Interface meant to release the resources hold by the object.
-    """
-
-    @abstractmethod
-    def close(self):
-        """Releases any resources associated to this object.
-        """
-        raise NotImplementedError
