+++ kernel/tools/attrdict.py
