+++ kernel/tools/dicts.py
