+++ kernel/tools/text.py
@@ -13,20 +13,6 @@
 #    License for the specific language governing permissions and limitations
 #    under the License.
 """
-import validators
-
-
-def is_url(string: str) -> bool:
-    """
-    :param string: The string to test.
-    :return: True if the string follows a URL format, False if not.
-    """
-    result = validators.url(string)
-
-    if isinstance(result, bool):
-        return result
-
-    return False
 
 
 class IndentedTextBuilder:
