+++ tests/kernel/intr/tools/test_sorting.py
@@ -15,8 +15,8 @@
 """
 from unittest import TestCase
 
-from cibyl.utils.sorting import (BubbleSortAlgorithm, NativeComparator, nsort,
-                                 sort)
+from kernel.tools.sorting import (BubbleSortAlgorithm, NativeComparator, nsort,
+                                  sort)
 
 
 class TestNativeComparator(TestCase):
